+++ .travis.yml
@@ -70,6 +70,8 @@ matrix:
     - env: TARGET=mips64-unknown-linux-gnuabi64
     - env: TARGET=mips-unknown-linux-gnu
     - env: TARGET=s390x-unknown-linux-gnu
+    - env: TARGET=asmjs-unknown-emscripten
+    - env: TARGET=wasm32-unknown-emscripten
 
     # beta
     - env: TARGET=x86_64-unknown-linux-gnu NO_ADD=1
