+++ ci/docker/wasm32-unknown-emscripten/node-wrapper.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+set -e
+
+me=$1
+shift
+dir=$(dirname $me)
+file=$(basename $me)
+
+cd $dir
+exec node $file "$@"
