+++ src/unix/notbsd/linux/musl/mod.rs
@@ -323,9 +323,7 @@ cfg_if! {
         pub use self::b64::*;
     } else if #[cfg(any(target_arch = "x86",
                         target_arch = "mips",
-                        target_arch = "arm",
-                        target_arch = "asmjs",
-                        target_arch = "wasm32"))] {
+                        target_arch = "arm"))] {
         mod b32;
         pub use self::b32::*;
     } else { }
